Fixed all warinings in DUMUX I encountered
- also added a CMake module to find ALBERTA (I'm not sure whether it's correct, though)
- added the files for the hysteresis example (forgot them last time so it didn't compile, sorry)


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dune-mux/trunk@689 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ dumux/operators/p1operatorextended.hh
@@ -853,7 +853,7 @@ namespace Dune
 //  	  std::cout << g.comm().rank() << ": " << "hanging nodes=" << hangingnodes << " links=" << links.size() << std::endl;
    
 	  // set size of all rows to zero
-	  for (size_t i=0; i<gv.indexSet().size(n); i++)
+	  for (size_t i=0; (int) i < (int) gv.indexSet().size(n); i++)
 		A.setrowsize(i,0); 
 
 	  // build needs a flag for all entities of all codims
